lexigramm/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.13.3
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
Django==4.2.19
django-bootstrap-v5==1.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lemminflect==0.2.3
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
nltk==3.9.1
numpy==2.2.3
packaging==24.2
preshed==3.0.9
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-docx==1.1.2
pywin32==308
regex==2024.11.6
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.1.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.4
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
